Move typechain to types folder
