+++ packages/safe-core-sdk/package.json
@@ -11,11 +11,11 @@
     "SDK"
   ],
   "scripts": {
-    "typechain-ethers": "typechain --target ethers-v5 --out-dir 'typechain/ethers-v5' '../../node_modules/@gnosis.pm/safe-contracts/build/contracts/**/*.json' '../../node_modules/openzeppelin-solidity/build/contracts/ERC20Mintable.json'",
-    "typechain-web3": "typechain --target web3-v1 --out-dir 'typechain/web3-v1' '../../node_modules/@gnosis.pm/safe-contracts/build/contracts/**/*.json' '../../node_modules/openzeppelin-solidity/build/contracts/ERC20Mintable.json'",
+    "typechain-ethers": "typechain --target ethers-v5 --out-dir 'src/types/typechain/ethers-v5' '../../node_modules/@gnosis.pm/safe-contracts/build/contracts/**/*.json' '../../node_modules/openzeppelin-solidity/build/contracts/ERC20Mintable.json'",
+    "typechain-web3": "typechain --target web3-v1 --out-dir 'src/types/typechain/web3-v1' '../../node_modules/@gnosis.pm/safe-contracts/build/contracts/**/*.json' '../../node_modules/openzeppelin-solidity/build/contracts/ERC20Mintable.json'",
     "typechain": "yarn typechain-ethers && yarn typechain-web3",
     "postinstall": "yarn typechain",
-    "unbuild": "rimraf dist artifacts cache typechain .nyc_output",
+    "unbuild": "rimraf dist artifacts cache .nyc_output",
     "build": "hardhat compile && tsc",
     "test:ganache:web3": "export TEST_NETWORK=ganache && export ETH_LIB=web3 && hardhat --network localhost deploy && nyc hardhat --network localhost test",
     "test:ganache:ethers": "export TEST_NETWORK=ganache && export ETH_LIB=ethers && hardhat --network localhost deploy && nyc hardhat --network localhost test",
