+++ packages/safe-core-sdk/src/contracts/MultiSend/MultiSendWeb3Contract.ts
@@ -1,4 +1,4 @@
-import { MultiSend } from '../../../typechain/web3-v1/MultiSend'
+import { MultiSend } from '../../types/typechain/web3-v1/MultiSend'
 import MultiSendContract from './MultiSendContract'
 
 class MultiSendWeb3Contract implements MultiSendContract {
