+++ packages/safe-core-sdk/src/ethereumLibs/EthersAdapter.ts
@@ -1,12 +1,12 @@
 import { TransactionResponse } from '@ethersproject/abstract-provider'
 import { Provider } from '@ethersproject/providers'
 import { BigNumber, Contract, Signer } from 'ethers'
-import { GnosisSafe__factory } from '../../typechain/ethers-v5/factories/GnosisSafe__factory'
-import { MultiSend__factory } from '../../typechain/ethers-v5/factories/MultiSend__factory'
 import { ContractNetworkConfig } from '../configuration/contracts'
 import GnosisSafeEthersV5Contract from '../contracts/GnosisSafe/GnosisSafeEthersV5Contract'
 import MultiSendEthersV5Contract from '../contracts/MultiSend/MultiSendEthersV5Contract'
-import { AbiItem } from '../utils/types'
+import { AbiItem } from '../types'
+import { GnosisSafe__factory } from '../types/typechain/ethers-v5/factories/GnosisSafe__factory'
+import { MultiSend__factory } from '../types/typechain/ethers-v5/factories/MultiSend__factory'
 import EthAdapter, { EthAdapterTransaction, GnosisSafeContracts } from './EthAdapter'
 
 export interface EthersAdapterConfig {
