+++ packages/safe-core-sdk/tests/utils/setupContracts.ts
@@ -6,7 +6,7 @@ import {
   GnosisSafe,
   MultiSend,
   SocialRecoveryModule
-} from '../../typechain/ethers-v5'
+} from '../../src/types/typechain/ethers-v5'
 
 export const getSafeSingleton = async () => {
   const SafeDeployment = await deployments.get('GnosisSafe')
